Refactor test to not duplicate order info
